refactor: move layout logic to separate component
